Updated README.md & minor refactoring
